requirements.txt
anthropic
python-dotenv
pandas
numpy
requests
pyyaml
duckdb
pybaseball
